requirements.txt
discord.py
psutil
pyperclip
pyautogui
opencv-python